bug with local names
sometimes local names are shorter.
back in the day, when I used powershell core, episode titles sometimes got doubled and were very long.

I think whisper get_writer has a limit on the file name, and truncates it.

have to use startswith instead of ==.

for all new episodes, because I'm using a Python funciton app, the titles should not get doubled and should not be this long